Changed the way Playlist entity is handled: the primary key is just an auto-incremental field. Easier to implement and resolves the problem with the generation of the id.

+++ app/controllers/Playlist.java
@@ -41,7 +41,7 @@ public class Playlist extends Controller
 	    }
 	    PlaylistForm playlist = playlistForm.get();
 	    
-	    int id = PersistPlaylist.savePlaylist(session(SESSION.USERNAME.toString()), playlist.getName());
+	    long id = PersistPlaylist.savePlaylist(session(SESSION.USERNAME.toString()), playlist.getName());
 	    
 		result.put("name", playlist.getName());
 		result.put("id", id);
